requirements.txt
streamlit
phidata
google-generativeai
yfinance
googlesearch-python
python-dotenv
pycountry
